+++ NEWS.md
@@ -1,8 +1,14 @@
+# democracyData 0.2.2
+
+* Now includes Anckar and Fredriksson's (2018) dataset of authoritarian regimes (`anckar`) and the Support Vector Machine index of democracy (two versions, Grundler and Krieger 2018 and 2016)
+
+* Fixed bugs in the `download_reign` function due to a change of schema in their google sheet.
+
 # democracyData 0.2.1
 
 * Now includes the Economist Intelligence Index's democracy index (`eiu`).
 
-* Updated the `bmr` dataset to the katest version (3.0), which includes a new variable, `democracy_femalesuffrage`.
+* Updated the `bmr` dataset to the latest version (3.0), which includes a new variable, `democracy_femalesuffrage`.
 
 * Fixed bugs with the `download_reign` function not picking up the right google spreadsheet.